CLONE_UNCONTAIN: Add a new flag for clone to control access to root PID name space
The CLONE_UNCONTAIN flag is required when creating a new PID name space
in order to grant the same space CAP_ADMIN_NS, else the capability is
stripped when the name space is created at do_fork() time.  This will
allow for complete backward compatibility with the security model in
use by name spaces and various container managers.

The container manager would be responsible for the setting up a name
space with CLONE_UNCONTAIN and dealing with all the
restrictions/permissions that come with this capability.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>